Challenge 1 - Justin/requirements.txt
torch==1.13.1
torchvision==0.14.1
Pillow==8.0.1
colorama==0.4.4